% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DetectIE.R
\name{DetectIE}
\alias{DetectIE}
\title{Detect Internet Explorer}
\usage{
DetectIE()
}
\arguments{
\item{NULL}{no inputs are required}
}
\value{
none, called for side effect
}
\description{
Use in a markdown file. Older version of IE will not render some of the charts or tables in knitr documents. The function inserts a html snippet which shows a warning message if the user is using Internet Explorer. The snippet uses VBScript code which only IE can interpret.  The html snippet will have no effect in Chrome as the browser will not interpret VBScript.
}
\examples{
```{r, echo=FALSE, results='asis'}
DetectIE()
```

}
\author{
Ryan Kuhn, CPA; \email{ryan.l.kuhn@questdiagnostics.com}
}

