% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGX.R
\docType{package}
\name{DGX}
\alias{DGX}
\alias{DGX-package}
\title{DGX: a package designed for Quest Diagnostics}
\description{
DGX provides useful functions for the Health Plan Finance function at Quest Diagnostics.
}
\author{
Ryan Kuhn, CPA; \email{ryan.l.kuhn@questdiagnostics.com}
}

