% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_Percent.R
\name{Print_Percent}
\alias{Print_Percent}
\title{Format a number as a percentage for printing}
\usage{
Print_Percent(x, decimals = 0)
}
\arguments{
\item{x}{a numeric or character value}

\item{decimals}{the number of decimal places to round to}
}
\description{
This function takes a character or numeric value as input. The value is returned as a character value with the percent sign appended
}
\author{
Ryan Kuhn, CPA; \email{ryan.l.kuhn@questdiagnostics.com}
}

