% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print_Thousands.R
\name{Print_Thousands}
\alias{Print_Thousands}
\title{Format a number for printing}
\usage{
Print_Thousands(x, boolcash = T, decimals = -3)
}
\arguments{
\item{x}{a numeric or character value}

\item{boolcash}{TRUE to prepend the return value with $}

\item{decimals}{the number of decimal places to round to}
}
\description{
This function takes a character or numeric value as input. The value is returned as a character value with the % appended
}
\author{
Ryan Kuhn, CPA; \email{ryan.l.kuhn@questdiagnostics.com}
}

