% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Read_QDSS.R
\name{Read_QDSS}
\alias{Read_QDSS}
\title{Read .csv from QDSS}
\usage{
Read_QDSS(x, cols = "character")
}
\arguments{
\item{x}{csv file downloaded from QDSS}

\item{cols}{character. A vector of classes to be assumed for the columns. Recycled as necessary}
}
\description{
Wrapper for read.csv file setup to read .csv files downloaded from QDSS.  The function sets fileEncoding option to UCS-2LE which is commonly forgotten for QDSS data.
}
\author{
Ryan Kuhn, CPA; \email{ryan.l.kuhn@questdiagnostics.com}
}

